Second pass at report generation. Refactored PostProcessor into Fixed and Variable Processors.
AllocationReport now works on single destination data set (whether a report of the main data). If a report data set, the data is copied before running the allocation process.
